# Conflicts:
#	StrongDC.vcxproj
#	StrongDC.vcxproj.filters
#	windows/TypedListViewCtrl.h
#	windows/resource.h

+++ airdcpp/airdcpp/modules/AutoSearch.cpp
@@ -1,5 +1,5 @@
 /*
-* Copyright (C) 2011-2017 AirDC++ Project
+* Copyright (C) 2011-2018 AirDC++ Project
 *
 * This program is free software; you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
@@ -360,9 +360,9 @@ bool AutoSearch::updateSearchTime() noexcept {
 		//add the start (or end) hours and minutes (if needed)
 		auto& timeStruct = toEnabled ? startTime : endTime;
 		if (timeStruct.hour > nextSearch.time_of_day().hours()) {
-			nextSearch += (hours(timeStruct.hour) + minutes(timeStruct.minute)) - (hours(nextSearch.time_of_day().hours()) + minutes(nextSearch.time_of_day().minutes()));
+			nextSearch += (hours(timeStruct.hour) + minutes(timeStruct.minute)) - (hours(static_cast<long>(nextSearch.time_of_day().hours())) + minutes(static_cast<long>(nextSearch.time_of_day().minutes())));
 		} else if ((timeStruct.hour == nextSearch.time_of_day().hours() && timeStruct.minute > nextSearch.time_of_day().minutes())) {
-			nextSearch += minutes(timeStruct.minute - nextSearch.time_of_day().minutes());
+			nextSearch += minutes(timeStruct.minute - static_cast<long>(nextSearch.time_of_day().minutes()));
 		}
 	};
